feat: Viewタイプごとに切り替えれるようにした、Simpleリストは切り出した
