+++ pages/kusa/contributions-by-event.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+type Props = {
+  result: any;
+};
+
+const ContributionsByEvent = (props: Props) => {
+  return <>Contributions By Event</>;
+};
+
+export default ContributionsByEvent;
