+++ pages/kusa/contributions-by-repo.tsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+type Props = {
+  result: any;
+};
+
+const ContributionsByRepo = (props: Props) => {
+  return <>Contributions By Repo</>;
+};
+
+export default ContributionsByRepo;
