+++ pages/kusa/contributions.tsx
@@ -1,7 +1,10 @@
 import React from 'react';
 import { useState } from 'react';
+import ContributionsByRepo from './contributions-by-repo';
+import ContributionsByEvent from './contributions-by-event';
+import ContributionsSimple from './contributions-simple';
 
-type PullRequestEventPayload = {
+export type PullRequestEventPayload = {
   action: string;
   number: number;
   pull_request: {
@@ -12,7 +15,7 @@ type PullRequestEventPayload = {
   };
 };
 
-type IssuesEventPayload = {
+export type IssuesEventPayload = {
   action: string;
   number: number;
   issue: {
@@ -23,34 +26,34 @@ type IssuesEventPayload = {
   };
 };
 
-type Commit = {
+export type Commit = {
   message: string;
   sha: string;
   url: string;
 };
 
-type PushEventPayload = {
+export type PushEventPayload = {
   before: string;
   head: string;
   ref: string;
   commits: Commit[];
 };
 
-type DeleteEventPayload = {
+export type DeleteEventPayload = {
   ref: string;
   ref_type: string;
 };
 
-type CreateEventPayload = {
+export type CreateEventPayload = {
   ref: string;
   ref_type: string;
 };
 
-type WatchEventPayload = {
+export type WatchEventPayload = {
   action: string;
 };
 
-type IssueCommentEventPayload = {
+export type IssueCommentEventPayload = {
   action: string;
   comment: {
     url: string;
@@ -64,7 +67,7 @@ type IssueCommentEventPayload = {
   };
 };
 
-type ForkEventPayload = {
+export type ForkEventPayload = {
   forkee: {
     url: string;
     html_url: string;
@@ -73,12 +76,12 @@ type ForkEventPayload = {
   };
 };
 
-type GitHubRepo = {
+export type GitHubRepo = {
   url: string;
   name: string;
 };
 
-type GitHubEventType =
+export type GitHubEventType =
   | 'PullRequestEvent'
   | 'IssuesEvent'
   | 'PushEvent'
@@ -88,7 +91,7 @@ type GitHubEventType =
   | 'IssueCommentEvent'
   | 'ForkEvent';
 
-type GitHubEvent = {
+export type GitHubEvent = {
   id: number;
   repo: GitHubRepo;
   created_at: string;
@@ -97,101 +100,8 @@ type GitHubEvent = {
   payload: any;
 };
 
-const toHtmlUrl = (url: string) => {
+export const toHtmlUrl = (url: string) => {
   return url.replace('api.github.com/repos', 'github.com');
-};
-const IssueEvent = ({ payload }: { payload: IssuesEventPayload }) => {
-  return (
-    <div>
-      <span className="text-blue-600 hover:underline">
-        <a href={payload.issue.html_url} target="_blank" rel="noreferrer">
-          #{payload.issue.number} {payload.issue.title}
-        </a>
-      </span>
-      {payload.action}
-    </div>
-  );
-};
-
-const PullRequestEvent = ({ payload }: { payload: PullRequestEventPayload }) => {
-  return (
-    <div>
-      <span className="text-blue-600 hover:underline">
-        <a href={payload.pull_request.html_url} target="_blank" rel="noreferrer">
-          #{payload.pull_request.number} {payload.pull_request.title}
-        </a>
-      </span>
-      {payload.action}
-    </div>
-  );
-};
-
-const PushEvent = ({ payload }: { payload: PushEventPayload }) => {
-  return (
-    <ul className="list-disc">
-      {payload.commits?.map((c) => {
-        return (
-          <li key={c.sha} className="ml-8">
-            <a
-              href={toHtmlUrl(c.url).replace('commits', 'commit')}
-              target="_blank"
-              rel="noreferrer"
-              className="text-blue-600 hover:underline"
-            >
-              {c.sha.substring(0, 6)}
-            </a>{' '}
-            {c.message}
-          </li>
-        );
-      })}
-    </ul>
-  );
-};
-
-const DeleteEvent = ({ payload }: { payload: DeleteEventPayload }) => {
-  return (
-    <div>
-      {payload.ref_type}: {payload.ref} Deleted
-    </div>
-  );
-};
-
-const CreateEvent = ({ payload }: { payload: CreateEventPayload }) => {
-  return (
-    <div>
-      {payload.ref_type}: {payload.ref} Created
-    </div>
-  );
-};
-
-const WatchEvent = ({ payload }: { payload: WatchEventPayload }) => {
-  return <div>{payload.action}</div>;
-};
-
-const IssueCommentEvent = ({ payload }: { payload: IssueCommentEventPayload }) => {
-  return (
-    <div>
-      <span className="text-blue-600 hover:underline">
-        <a href={payload.comment.html_url} target="_blank" rel="noreferrer">
-          #{payload.issue.number} {payload.issue.title}
-        </a>
-      </span>
-      <span className="ml-1">comment {payload.action}</span>
-    </div>
-  );
-};
-
-const ForkEvent = ({ payload }: { payload: ForkEventPayload }) => {
-  return (
-    <div>
-      Fork to
-      <span className="text-blue-600 hover:underline">
-        <a href={payload.forkee.html_url} target="_blank" rel="noreferrer">
-          {payload.forkee.full_name}
-        </a>
-      </span>
-    </div>
-  );
 };
 
 type Props = {
@@ -200,16 +110,13 @@ type Props = {
 };
 
 const Contributions = (props: Props) => {
-  const [open, setOpen] = useState<boolean>();
   const [selectedView, setSelectedView] = useState<string>('simple');
   const views = [
     { id: 'simple', name: 'Simple List' },
     { id: 'repo', name: 'Group By Repo' },
+    { id: 'event', name: 'Group By Event' },
   ];
 
-  const handleSummaryOpen = () => {
-    setOpen((prev) => !prev);
-  };
   return (
     <div>
       <h2 className="font-bold col-start-1 col-end-4">Recent {props.user} Events</h2>
@@ -233,45 +140,10 @@ const Contributions = (props: Props) => {
           );
         })}
       </nav>
-      <div className="grid grid-cols-10">
-        <button
-          onClick={handleSummaryOpen}
-          className="col-start-5 col-end-10 bg-white hover:bg-gray-100 text-gray-800 font-semibold py-1 px-2 border border-gray-400 rounded shadow"
-        >
-          {open ? 'Fold up' : 'Open'} All Details
-        </button>
-      </div>
-      <div>
-        {props.result.map((row: GitHubEvent) => {
-          return (
-            <div key={row.id} className="grid grid-cols-10 gap-4">
-              <div className="col-start-1 col-end-1">{row.created_at.split('T')[0]}</div>
-              <div className="col-start-2 col-end-4 whitespace-nowrap text-blue-600 hover:underline">
-                {row.repo.url ? (
-                  <a href={toHtmlUrl(row.repo.url)} target="_blank" rel="noreferrer">
-                    {row.repo.name}
-                  </a>
-                ) : (
-                  <>{row.repo.name}</>
-                )}
-              </div>
-              <div className="col-start-5 col-end-10">
-                <details open={open}>
-                  <summary>{row.type}</summary>
-                  {row.type === 'PullRequestEvent' && <PullRequestEvent payload={row.payload}></PullRequestEvent>}
-                  {row.type === 'PushEvent' && <PushEvent payload={row.payload}></PushEvent>}
-                  {row.type === 'IssuesEvent' && <IssueEvent payload={row.payload}></IssueEvent>}
-                  {row.type === 'DeleteEvent' && <DeleteEvent payload={row.payload}></DeleteEvent>}
-                  {row.type === 'CreateEvent' && <CreateEvent payload={row.payload}></CreateEvent>}
-                  {row.type === 'WatchEvent' && <WatchEvent payload={row.payload}></WatchEvent>}
-                  {row.type === 'IssueCommentEvent' && <IssueCommentEvent payload={row.payload}></IssueCommentEvent>}
-                  {row.type === 'ForkEvent' && <ForkEvent payload={row.payload}></ForkEvent>}
-                </details>
-              </div>
-            </div>
-          );
-        })}
-      </div>
+
+      {selectedView == 'simple' && <ContributionsSimple result={props.result}></ContributionsSimple>}
+      {selectedView == 'repo' && <ContributionsByRepo result={props.result}></ContributionsByRepo>}
+      {selectedView == 'event' && <ContributionsByEvent result={props.result}></ContributionsByEvent>}
     </div>
   );
 };
